move PlaceholderResolver
